move EncryptedMessage class to isolated file
