+++ transport/src/main/java/me/uport/sdk/transport/Crypto.kt
@@ -1,39 +1,15 @@
 package me.uport.sdk.transport
 
-import kotlinx.serialization.SerialName
-import kotlinx.serialization.Serializable
-import kotlinx.serialization.json.JSON
 import me.uport.knacl.nacl
 import me.uport.sdk.core.decodeBase64
 import me.uport.sdk.core.toBase64
+import me.uport.sdk.transport.EncryptedMessage.Companion.ASYNC_ENC_ALGORITHM
 
 /**
  * This class exposes methods to encrypt and decrypt messages according to the uPort spec at
  * https://github.com/uport-project/specs/blob/develop/messages/encryption.md
  */
 object Crypto {
-
-    /**
-     * This class encapsulates an encrypted message that was produced using
-     * https://github.com/uport-project/specs/blob/develop/messages/encryption.md
-     */
-    @Serializable
-    data class EncryptedMessage(
-            @SerialName("version")
-            val version: String = ASYNC_ENC_ALGORITHM,
-            @SerialName("nonce")
-            val nonce: String,
-            @SerialName("ephemPublicKey")
-            val ephemPublicKey: String,
-            @SerialName("ciphertext")
-            val ciphertext: String
-    ) {
-        fun toJson(): String = JSON.stringify(EncryptedMessage.serializer(), this)
-
-        companion object {
-            fun fromJson(json: String): EncryptedMessage? = JSON.nonstrict.parse(EncryptedMessage.serializer(), json)
-        }
-    }
 
     /**
      * Calculates the publicKey usable for encryption corresponding to the given [secretKey]
@@ -81,7 +57,5 @@ object Crypto {
                 secretKey) ?: throw RuntimeException("Could not decrypt message")
         return decrypted.toString(Charsets.UTF_8).unpad()
     }
-
-    private const val ASYNC_ENC_ALGORITHM = "x25519-xsalsa20-poly1305"
 
 }
